# Optimize by minimizing len() calls